# openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FastAPI", "version": "0.1.0" },
  "paths": {
...
    "/v1/products": {
      "get": {
        "summary": "List Products",
        "operationId": "list_products_v1_products_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Product",
        "operationId": "create_product_v1_products_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_product_v1_products_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/products/{product_id}": {
      "post": {
        "summary": "Update Product",
        "operationId": "update_product_v1_products__product_id__post",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Product Id" }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_product_v1_products__product_id__post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Product",
        "operationId": "retrieve_product_v1_products__product_id__get",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Product Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Product" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Product",
        "operationId": "delete_product_v1_products__product_id__delete",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Product Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProductDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
    "/v1/prices": {
      "get": {
        "summary": "List Prices",
        "operationId": "list_prices_v1_prices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PriceList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Price",
        "operationId": "create_price_v1_prices_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_price_v1_prices_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Price" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/v1/prices/{price_id}": {
      "post": {
        "summary": "Update Price",
        "operationId": "update_price_v1_prices__price_id__post",
        "parameters": [
          {
            "name": "price_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Price Id" }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Body_update_price_v1_prices__price_id__post"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Price" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Retrieve Price",
        "operationId": "retrieve_price_v1_prices__price_id__get",
        "parameters": [
          {
            "name": "price_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Price Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Price" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Price",
        "operationId": "delete_price_v1_prices__price_id__delete",
        "parameters": [
          {
            "name": "price_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Price Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PriceDelete" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_price_v1_prices_post": {
        "properties": {
          "active": { "type": "string", "title": "Active" },
          "currency": { "type": "string", "title": "Currency" },
          "price_type": { "type": "string", "title": "Price Type" },
          "unit_amount": { "type": "integer", "title": "Unit Amount" },
          "default": { "type": "string", "title": "Default" },
          "variant": { "type": "string", "title": "Variant" },
          "customer_unit_amount_maximum": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Maximum"
          },
          "customer_unit_amount_minimum": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Minimum"
          },
          "customer_unit_amount_preset": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Preset"
          },
          "recurring_interval": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Recurring Interval"
          },
          "recurring_interval_count": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Recurring Interval Count"
          }
        },
        "type": "object",
        "required": [
          "active",
          "currency",
          "price_type",
          "unit_amount",
          "default",
          "variant"
        ],
        "title": "Body_create_price_v1_prices_post"
      },
      "Body_create_product_v1_products_post": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "active": { "type": "string", "title": "Active" },
          "shippable": { "type": "string", "title": "Shippable" },
          "preorder": { "type": "string", "title": "Preorder" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["name", "active", "shippable", "preorder"],
        "title": "Body_create_product_v1_products_post"
      },
...
      "Body_update_price_v1_prices__price_id__post": {
        "properties": {
          "active": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Active"
          },
          "default": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Default"
          },
          "customer_unit_amount_maximum": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Maximum"
          },
          "customer_unit_amount_minimum": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Minimum"
          },
          "customer_unit_amount_preset": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Customer Unit Amount Preset"
          },
          "recurring_interval": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Recurring Interval"
          },
          "recurring_interval_count": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Recurring Interval Count"
          }
        },
        "type": "object",
        "title": "Body_update_price_v1_prices__price_id__post"
      },
      "Body_update_product_v1_products__product_id__post": {
        "properties": {
          "name": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Name"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "active": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Active"
          },
          "shippable": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Shippable"
          },
          "preorder": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Preorder"
          }
        },
        "type": "object",
        "title": "Body_update_product_v1_products__product_id__post"
      },
...
      "CustomerUnitAmount": {
        "properties": {
          "maximum": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Maximum"
          },
          "minimum": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Minimum"
          },
          "preset": { "type": "integer", "title": "Preset", "default": 1 }
        },
        "type": "object",
        "title": "CustomerUnitAmount"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Price": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "active": { "type": "boolean", "title": "Active" },
          "currency": { "type": "string", "title": "Currency" },
          "type": { "$ref": "#/components/schemas/PriceTypeEnum" },
          "unit_amount": { "type": "integer", "title": "Unit Amount" },
          "default": { "type": "boolean", "title": "Default" },
          "customer_unit_amount": {
            "anyOf": [
              { "$ref": "#/components/schemas/CustomerUnitAmount" },
              { "type": "null" }
            ]
          },
          "recurring": {
            "anyOf": [
              { "$ref": "#/components/schemas/Recurring" },
              { "type": "null" }
            ]
          },
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "price" }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "active",
          "currency",
          "type",
          "unit_amount",
          "default",
          "customer_unit_amount",
          "recurring",
          "id"
        ],
        "title": "Price"
      },
      "PriceDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": { "type": "string", "title": "Object", "default": "price" },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "PriceDelete"
      },
      "PriceList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": { "type": "string", "title": "Url", "default": "/v1/prices" },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Price" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["has_more"],
        "title": "PriceList"
      },
      "PriceTypeEnum": {
        "type": "string",
        "enum": ["one_time", "subscription"],
        "title": "PriceTypeEnum"
      },
      "Product": {
        "properties": {
          "created": { "type": "integer", "title": "Created" },
          "updated": { "type": "integer", "title": "Updated" },
          "livemode": { "type": "boolean", "title": "Livemode" },
          "name": { "type": "string", "title": "Name" },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          },
          "images": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Images",
            "default": []
          },
          "active": { "type": "boolean", "title": "Active" },
          "shippable": { "type": "boolean", "title": "Shippable" },
          "preorder": { "type": "boolean", "title": "Preorder" },
          "id": { "type": "string", "title": "Id" },
          "handle": { "type": "string", "title": "Handle" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "product"
          }
        },
        "type": "object",
        "required": [
          "created",
          "updated",
          "livemode",
          "name",
          "active",
          "shippable",
          "preorder",
          "id",
          "handle"
        ],
        "title": "Product"
      },
      "ProductDelete": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "object": {
            "type": "string",
            "title": "Object",
            "default": "product"
          },
          "deleted": { "type": "boolean", "title": "Deleted" }
        },
        "type": "object",
        "required": ["id", "deleted"],
        "title": "ProductDelete"
      },
      "ProductList": {
        "properties": {
          "object": { "type": "string", "title": "Object", "default": "list" },
          "url": {
            "type": "string",
            "title": "Url",
            "default": "/v1/products"
          },
          "has_more": { "type": "boolean", "title": "Has More" },
          "data": {
            "items": { "$ref": "#/components/schemas/Product" },
            "type": "array",
            "title": "Data",
            "default": []
          }
        },
        "type": "object",
        "required": ["has_more"],
        "title": "ProductList"
      },
      "Recurring": {
        "properties": {
          "interval": { "$ref": "#/components/schemas/RecurringTypeEnum" },
          "interval_count": { "type": "integer", "title": "Interval Count" }
        },
        "type": "object",
        "required": ["interval", "interval_count"],
        "title": "Recurring"
      },
      "RecurringTypeEnum": {
        "type": "string",
        "enum": ["day", "week", "month", "year"],
        "title": "RecurringTypeEnum"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    }
  }
}